+++ source/dem/dem.cc
@@ -946,7 +946,7 @@ DEMSolver<dim>::set_insertion_type(const DEMSolverParameters<dim> &parameters)
                                              distribution_object_container);
     }
   else if (parameters.insertion_info.insertion_method ==
-           Parameters::Lagrangian::InsertionInfo::InsertionMethod::from_file)
+           Parameters::Lagrangian::InsertionInfo::InsertionMethod::file)
     {
       insertion_object =
         std::make_shared<FromFileInsertion<dim>>(parameters,